Added GitMixin
